refactor: split home_screen into smaller widget components
Extract presentational pieces of the ~880-line home screen into reusable
widgets under lib/view/widgets/: badge_text_input_field,
badge_clipart_picker, badge_control_tab_bar, badge_control_tab_view, and
badge_action_buttons. Save/transfer business logic stays in the screen and
is passed in as callbacks. No behavior change.

Part of #1852. Closes #1857.
